requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
dblpy==0.4.0
discord==1.0.1
discord-py-slash-command==1.0.8.5
discord.py==1.6.0
disputils==0.1.3
dpymenus==1.3.1
emoji==0.6.0
Flask==1.1.2
Flighter==2.0.4
humanize==3.2.0
jishaku==1.20.0.220
numpy==1.19.5
Pillow==8.1.0
qrcode==6.1
urllib3==1.26.3
